Adding unit tests to increase coverage. Combine two music_pet test files.
