+++ tests/test_music_pet.py
@@ -2,8 +2,6 @@
 import unittest
 import os
 
-from music_pet import meta
-
 
 class TestOnParseCUEFile(unittest.TestCase):
 
@@ -11,6 +9,8 @@ class TestOnParseCUEFile(unittest.TestCase):
         return
 
     def test_single_flac_CUE_utf8(self):
+        from music_pet import meta
+
         INPUT_FILE = u"testfiles/CUETestFile1.utf8.cue"
 
         album_list = meta.parse_cue(INPUT_FILE)
@@ -77,6 +77,8 @@ class TestOnParseCUEFile(unittest.TestCase):
         self.assertEqual(albums, albums_expected)
 
     def test_multitrack_flac_CUE_utf8(self):
+        from music_pet import meta
+
         INPUT_FILE = u"testfiles/CUETestFile2.utf8.cue"
 
         album_list = meta.parse_cue(INPUT_FILE)
@@ -157,6 +159,8 @@ class TestOnParseCUEFile(unittest.TestCase):
 
 
     def test_INI_utf8(self):
+        from music_pet import meta
+
         CUE_FILE = u"testfiles/CUETestFile2.utf8.cue"
         INPUT_FILE = u"testfiles/INITestFile1.ini"
 
@@ -250,3 +254,330 @@ class TestOnParseCUEFile(unittest.TestCase):
         self.assertEqual(albums, albums_expected)
 
 
+class UnitTest_music_pet__meta__Track(unittest.TestCase):
+
+    def setUp(self):
+        return
+
+    def tearDown(self):
+        return
+
+    def test_tracknumber_1(self):
+        from music_pet.meta import Track
+
+        t = Track()
+        t.tracknumber = 1
+        self.assertEqual(t.tracknumber, u"01")
+
+    def test_tracknumber_2(self):
+        from music_pet.meta import Track
+
+        t = Track()
+        t.tracknumber = 14
+        self.assertEqual(t.tracknumber, u"14")
+
+    def test_tracknumber_4(self):
+        from music_pet.meta import Track
+
+        t = Track()
+        t.tracknumber = 143
+        self.assertEqual(t.tracknumber, u"143")
+
+    def test_tracknumber_3(self):
+        from music_pet.meta import Track
+
+        t = Track()
+        t.totaltracks = 1024
+        t.tracknumber = 14
+        self.assertEqual(t.tracknumber, u"0014")
+
+    def test_tracknumber_2(self):
+        from music_pet.meta import Track
+
+        t = Track()
+        t.tracknumber = 14
+        t.totaltracks = 1024
+        t.refresh_tracknumber()
+        self.assertEqual(t.tracknumber, u"0014")
+
+
+class UnitTest_music_pet__audio__flac__FLAC(unittest.TestCase):
+
+    def setUp(self):
+        return
+
+    def tearDown(self):
+        return
+
+    def test_set_next_start_time_from_album(self):
+        from music_pet.meta import Track, Album
+        from music_pet.audio.flac import FLAC
+
+        t1 = Track()
+        t1.tracknumber = 1
+        t1.set_tag(u"index_00", u"00:00:00")
+        t1.set_tag(u"original_file", u"CDImage.flac")
+        t2 = Track()
+        t2.tracknumber = 2
+        t2.set_tag(u"index_00", u"00:03:01")
+        t2.set_tag(u"original_file", u"CDImage.flac")
+        a = Album(tracks=[t1, t2])
+
+        f = FLAC(t1.data)
+        f.set_next_start_time_from_album(a)
+
+        self.assertTrue(f.has_tag(u"@time_to"))
+        print(f.metadata.detail())
+
+
+class UnitTest_music_pet__utils__path_from_pattern(unittest.TestCase):
+
+    def setUp(self):
+        return
+
+    def tearDown(self):
+        return
+
+    def test_1(self):
+        from music_pet.utils import path_from_pattern
+
+        PATTERN = u"/Users/normaluser/music_output/<%(prefix)s >%(album)s< (%(suffix)s)>/<<%(discnumber)s->%(tracknumber)s >%(title)s.flac"
+        D = {
+            u"album": u"看不懂到底有什么想法",
+            u"title": u"有没有搞错",
+        }
+
+        path = path_from_pattern(PATTERN, D)
+        path_expected = u"/Users/normaluser/music_output/看不懂到底有什么想法/有没有搞错.flac"
+
+        print(path)
+        self.assertEqual(path, path_expected, "The result is not correct!")
+
+    def test_2(self):
+        from music_pet.utils import path_from_pattern
+
+        PATTERN = u"/Users/normaluser/music_output/<%(prefix)s >%(album)s< (%(suffix)s)>/<<%(discnumber)s->%(tracknumber)s >%(title)s.flac"
+        D = {
+            u"album": u"看不懂到底有什么想法",
+            u"title": u"有没有搞错",
+            u"tracknumber": u"14",
+        }
+
+        path = path_from_pattern(PATTERN, D)
+        path_expected = u"/Users/normaluser/music_output/看不懂到底有什么想法/14 有没有搞错.flac"
+
+        print(path)
+        self.assertEqual(path, path_expected, "The result is not correct")
+
+    def test_3(self):
+        from music_pet.utils import path_from_pattern
+
+        PATTERN = u"/Users/normaluser/music_output/<%(prefix)s >%(album)s< (%(suffix)s)>/<<%(discnumber)s->%(tracknumber)s >%(title)s.flac"
+        D = {
+            u"album": u"看不懂到底有什么想法",
+            u"title": u"有没有搞错",
+            u"tracknumber": u"14",
+            u"discnumber": u"3",
+        }
+
+        path = path_from_pattern(PATTERN, D)
+        path_expected = u"/Users/normaluser/music_output/看不懂到底有什么想法/3-14 有没有搞错.flac"
+
+        print(path)
+        self.assertEqual(path, path_expected, "The result is not correct")
+
+    def test_4(self):
+        from music_pet.utils import path_from_pattern
+
+        PATTERN = u"/Users/normaluser/music_output/<%(prefix)s >%(album)s< (%(suffix)s)>/<<%(discnumber)s->%(tracknumber)s >%(title)s.flac"
+        D = {
+            u"album": u"看不懂到底有什么想法",
+            u"title": u"有没有搞错",
+            u"discnumber": u"3",
+        }
+
+        path = path_from_pattern(PATTERN, D)
+        path_expected = u"/Users/normaluser/music_output/看不懂到底有什么想法/有没有搞错.flac"
+
+        print(path)
+        self.assertEqual(path, path_expected, "The result is not correct")
+
+    def test_5(self):
+        from music_pet.utils import path_from_pattern
+
+        PATTERN = u"/Users/normaluser/music_output/<%(prefix)s >%(album)s< (%(suffix)s)>/<<%(discnumber)s->%(tracknumber)s >%(title)s.flac"
+        D = {
+            u"album": u"看不懂到底有什么想法",
+            u"title": u"有没有搞错",
+            u"tracknumber": u"14",
+            u"discnumber": u"3",
+            u"prefix": u"精选集",
+        }
+
+        path = path_from_pattern(PATTERN, D)
+        path_expected = u"/Users/normaluser/music_output/精选集 看不懂到底有什么想法/3-14 有没有搞错.flac"
+
+        print(path)
+        self.assertEqual(path, path_expected, "The result is not correct")
+
+    def test_6(self):
+        from music_pet.utils import path_from_pattern
+
+        PATTERN = u"/Users/normaluser/music_output/<%(prefix)s >%(album)s< (%(suffix)s)>/<<%(discnumber)s->%(tracknumber)s >%(title)s.flac"
+        D = {
+            u"album": u"看不懂到底有什么想法",
+            u"title": u"有没有搞错",
+            u"tracknumber": u"14",
+            u"discnumber": u"3",
+            u"prefix": u"精选集",
+            u"suffix": u"限量版",
+        }
+
+        path = path_from_pattern(PATTERN, D)
+        path_expected = u"/Users/normaluser/music_output/精选集 看不懂到底有什么想法 (限量版)/3-14 有没有搞错.flac"
+
+        print(path)
+        self.assertEqual(path, path_expected, "The result is not correct")
+
+
+class UnitTest_music_pet__utils__trim_quote(unittest.TestCase):
+
+    def setUp(self):
+        return
+
+    def tearDown(self):
+        return
+
+    def test_have_quote(self):
+        from music_pet.utils import trim_quote
+
+        input_string = u'''"Hello!"'''
+
+        output_string = trim_quote(input_string)
+        output_expected = u'''Hello!'''
+
+        self.assertEqual(output_string, output_expected)
+
+    def test_no_quote(self):
+        from music_pet.utils import trim_quote
+
+        input_string = u'''Hello!'''
+
+        output_string = trim_quote(input_string)
+        output_expected = u'''Hello!'''
+
+        self.assertEqual(output_string, output_expected)
+
+    def test_have_one_quote(self):
+        from music_pet.utils import trim_quote
+
+        input_string = u'''"Hello!'''
+
+        output_string = trim_quote(input_string)
+        output_expected = u'''"Hello!'''
+
+        self.assertEqual(output_string, output_expected)
+
+
+class UnitTest_music_pet__utils__remove_bom(unittest.TestCase):
+
+    def setUp(self):
+        import os
+
+        self.str_with_bom = "\xef\xbb\xbf1234567"
+        self.str_without_bom = "1234567"
+
+        try:
+            os.makedirs(u"tmp")
+        except:
+            pass
+
+        self.testfiles = [
+            u"tmp/file_with_bom.txt",
+            u"tmp/file_without_bom.txt",
+        ]
+
+        with open(u"tmp/file_with_bom.txt", "w") as fp:
+            fp.write(self.str_with_bom)
+        with open(u"tmp/file_without_bom.txt", "w") as fp:
+            fp.write(self.str_without_bom)
+
+        return
+
+    def tearDown(self):
+        import os
+
+        for f in self.testfiles:
+            os.remove(f)
+        return
+
+    def test_remove_bom(self):
+        from music_pet.utils import remove_bom
+
+        self.testfiles.append(u"tmp/file_with_bom_removed.txt")
+        remove_bom(u"tmp/file_with_bom.txt", u"tmp/file_with_bom_removed.txt")
+
+        with open(u"tmp/file_with_bom_removed.txt") as fp:
+            content = fp.read()
+
+        self.assertEqual(content, self.str_without_bom)
+
+    def test_remove_bom_not_exists(self):
+        from music_pet.utils import remove_bom
+
+        with self.assertRaises(ValueError):
+            remove_bom(u"tmp/file_without_bom.txt", u"tmp/file_with_bom_removed.txt")
+
+
+class UnitTest_music_pet__utils__iconv_file(unittest.TestCase):
+
+    def setUp(self):
+        return
+
+    def tearDown(self):
+        return
+
+
+class UnitTest_music_pet__utils__filename_safe(unittest.TestCase):
+
+    def setUp(self):
+        return
+
+    def tearDown(self):
+        return
+
+
+class UnitTest_music_pet__utils__ensure_parent_folder(unittest.TestCase):
+
+    def setUp(self):
+        return
+
+    def tearDown(self):
+        return
+
+
+class UnitTest_music_pet__utils__cli_escape(unittest.TestCase):
+
+    def setUp(self):
+        return
+
+    def tearDown(self):
+        return
+
+
+class UnitTest_music_pet__utils__parent_folder(unittest.TestCase):
+
+    def setUp(self):
+        return
+
+    def tearDown(self):
+        return
+
+
+class UnitTest_music_pet__utils__copy_to(unittest.TestCase):
+
+    def setUp(self):
+        return
+
+    def tearDown(self):
+        return
